+++ onlyoffice_odoo_documents/static/src/documents_view/onlyoffice_odoo_documents_controller_mixin.js
@@ -15,6 +15,7 @@ export const OnlyofficeDocumentsControllerMixin = {
         this.dialogService.add(CreateDialog, {
             folderId: this.env.searchModel.getSelectedFolderId(),
             context: this.props.context,
+            model: this.env.model
         });
     },
 };